Modified to be able to build on Windows. Modified not to use intptr_t as a variable of SG_INT_VALUE. Modified cache directory structure. Modified not to abort when reading cache process found invalid cache tag.

+++ src/bignum.c
@@ -593,7 +593,7 @@ int Sg_BignumBitSize(SgBignum *b)
   ASSERT(last >= 0);
   ASSERT(b->elements[last]);
   bitsize = WORD_BITS * last;
-  return bitsize + WORD_BITS - nlz((intptr_t)b->elements[last]);
+  return bitsize + WORD_BITS - nlz((long)b->elements[last]);
 }
 
 int Sg_BignumFirstBitSet(SgBignum *b)
