+++ src/sagittarius/sagittariusdefs.h
@@ -355,9 +355,9 @@ typedef enum  {
 
 /* fixnum */
 #define SG_INTP(obj)       (SG_TAG2(obj) == 1)
-#define SG_INT_VALUE(obj)  (((signed long int)SG_WORD(obj)) >> 2)
-#define SG_MAKE_INT(obj)   SG_OBJ(((intptr_t)(obj) << 2) + 1)
-#define SG_UINTP(obj)      (SG_INTP(obj) && ((signed long int)SG_WORD(obj) >= 0))
+#define SG_INT_VALUE(obj)  (((long)SG_WORD(obj)) >> 2)
+#define SG_MAKE_INT(obj)   SG_OBJ(((long)(obj) << 2) + 1)
+#define SG_UINTP(obj)      (SG_INTP(obj)&&((long)SG_WORD(obj) >= 0))
 #define SG_INT_SIZE        (SIZEOF_LONG * 8 - 3)
 #define SG_INT_MAX         ((1L << SG_INT_SIZE) - 1)
 #define SG_INT_MIN         (-SG_INT_MAX - 1)
@@ -365,7 +365,7 @@ typedef enum  {
 #define SG_CHAR(obj)       ((SgChar)(obj))
 #define SG_CHARP(obj)      (SG_TAG8(obj) == 3)
 #define SG_CHAR_VALUE(obj) SG_CHAR(((unsigned long)SG_WORD(obj)) >> 8)
-#define SG_MAKE_CHAR(obj)  SG_OBJ(((uintptr_t)(obj) << 8) + 0x03)
+#define SG_MAKE_CHAR(obj)  SG_OBJ(((unsigned long)(obj) << 8) + 0x03)
 /* SgChar is typedef of int32_t, so max value is 24 bits  */
 #define SG_CHAR_MAX        (0xffffff)
